🔥 refactor: 카테고리 패키지 위치 변경
