+++ storage/db-core/src/test/java/com/fstuckint/baedalyogieats/storage/db/core/store/CategoryRepositoryIT.java
@@ -1,4 +1,4 @@
-package com.fstuckint.baedalyogieats.storage.db.core.category;
+package com.fstuckint.baedalyogieats.storage.db.core.store;
 
 import static org.assertj.core.api.Assertions.assertThat;
